+++ frontend/src/hooks/useCreateTicket.tsx
@@ -59,7 +59,8 @@ const useCreateTicket = () => {
             );
             return data;
         } catch (error: any) {
-            toast.error(`${error?.response?.error}`);
+            console.error("create ticket error", error)
+            toast.error(`${error?.response?.data?.error}`);
         } finally {
             setLoading(false);
         }