+++ server/controllers/department/hhp/technicians/add_hhp_task.js
@@ -49,7 +49,30 @@ const AddHHPTask = async (req, res) => {
         requires_backup,
         rs_warranty,
     } = req.body;
-
+    console.log(
+        service_order_no,
+        date_booked,
+        model,
+        warranty,
+        engineer,
+        fault,
+        imei,
+        serial_number,
+        status,
+        ticket_number,
+        department,
+        job_added_by,
+        stores,
+        repairshopr_job_id,
+        repeat_repair,
+        created_at,
+        additional_info,
+        repairshopr_customer_id,
+        job_repair_no,
+        accessories_and_condition,
+        requires_backup,
+        rs_warranty
+    );
     try {
         await AddHHPTaskSchema.validate(req.body, { abortEarly: false });
         const findIfExists = await pool.query(
@@ -64,7 +87,7 @@ const AddHHPTask = async (req, res) => {
             });
         } else {
             const { rows } = await pool.query(
-                "INSERT INTO technician_tasks (service_order_no, date_booked, model, warranty, engineer, fault, imei, serial_number, unit_status, ticket_number, department, job_added_by, stores, repairshopr_job_id, repeat_repair, created_at, additional_info, repairshopr_customer_id) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22) returning id",
+                "INSERT INTO technician_tasks (service_order_no, date_booked, model, warranty, engineer, fault, imei, serial_number, unit_status, ticket_number, department, job_added_by, stores, repairshopr_job_id, repeat_repair, created_at, additional_info, repairshopr_customer_id) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18) returning id",
                 [
                     service_order_no,
                     date_booked,
@@ -84,12 +107,9 @@ const AddHHPTask = async (req, res) => {
                     created_at,
                     additional_info,
                     repairshopr_customer_id,
-                    job_repair_no,
-                    accessories_and_condition,
-                    requires_backup,
-                    rs_warranty,
                 ]
             );
+
             const fetchResult = await pool.query(returnDataWithNewRow, [
                 rows[0].id,
             ]);