+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.14] - 2025-01-19
+
+
+### Fixed
+
+-   increase client_max_body_size  in the nginx server file
+-   upload in book from so
+-   warranty codes and backup codes from rs
+
+
 ## [0.1.13] - 2025-01-18
 
 ### Added
@@ -12,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   store backup code when adding ticket and after creation
 -   store warranty code when adding ticket and after creation
 -   store job_repair_no code when adding ticket and after creation
--
 
 ### Fixed